+++ accelbyte_py_sdk/api/eventlog/operations/event/get_event_by_event_type_8d579c.py
@@ -61,11 +61,11 @@ class GetEventByEventTypeAndEventIDHandler(Operation):
 
         namespace: (namespace) REQUIRED str in path
 
-        offset: (offset) OPTIONAL float in query
+        offset: (offset) OPTIONAL int in query
 
         end_date: (endDate) REQUIRED str in query
 
-        page_size: (pageSize) REQUIRED float in query
+        page_size: (pageSize) REQUIRED int in query
 
         start_date: (startDate) REQUIRED str in query
 
@@ -95,9 +95,9 @@ class GetEventByEventTypeAndEventIDHandler(Operation):
     event_id: float                                                                                # REQUIRED in [path]
     event_type: float                                                                              # REQUIRED in [path]
     namespace: str                                                                                 # REQUIRED in [path]
-    offset: float                                                                                  # OPTIONAL in [query]
+    offset: int                                                                                    # OPTIONAL in [query]
     end_date: str                                                                                  # REQUIRED in [query]
-    page_size: float                                                                               # REQUIRED in [query]
+    page_size: int                                                                                 # REQUIRED in [query]
     start_date: str                                                                                # REQUIRED in [query]
 
     # endregion fields
@@ -218,7 +218,7 @@ class GetEventByEventTypeAndEventIDHandler(Operation):
         self.namespace = value
         return self
 
-    def with_offset(self, value: float) -> GetEventByEventTypeAndEventIDHandler:
+    def with_offset(self, value: int) -> GetEventByEventTypeAndEventIDHandler:
         self.offset = value
         return self
 
@@ -226,7 +226,7 @@ class GetEventByEventTypeAndEventIDHandler(Operation):
         self.end_date = value
         return self
 
-    def with_page_size(self, value: float) -> GetEventByEventTypeAndEventIDHandler:
+    def with_page_size(self, value: int) -> GetEventByEventTypeAndEventIDHandler:
         self.page_size = value
         return self
 
@@ -253,17 +253,17 @@ class GetEventByEventTypeAndEventIDHandler(Operation):
         elif include_empty:
             result["namespace"] = str()
         if hasattr(self, "offset") and self.offset:
-            result["offset"] = float(self.offset)
+            result["offset"] = int(self.offset)
         elif include_empty:
-            result["offset"] = float()
+            result["offset"] = int()
         if hasattr(self, "end_date") and self.end_date:
             result["endDate"] = str(self.end_date)
         elif include_empty:
             result["endDate"] = str()
         if hasattr(self, "page_size") and self.page_size:
-            result["pageSize"] = float(self.page_size)
+            result["pageSize"] = int(self.page_size)
         elif include_empty:
-            result["pageSize"] = float()
+            result["pageSize"] = int()
         if hasattr(self, "start_date") and self.start_date:
             result["startDate"] = str(self.start_date)
         elif include_empty:
@@ -327,9 +327,9 @@ class GetEventByEventTypeAndEventIDHandler(Operation):
         event_type: float,
         namespace: str,
         end_date: str,
-        page_size: float,
+        page_size: int,
         start_date: str,
-        offset: Optional[float] = None,
+        offset: Optional[int] = None,
     ) -> GetEventByEventTypeAndEventIDHandler:
         instance = cls()
         instance.event_id = event_id
@@ -358,17 +358,17 @@ class GetEventByEventTypeAndEventIDHandler(Operation):
         elif include_empty:
             instance.namespace = str()
         if "offset" in dict_ and dict_["offset"] is not None:
-            instance.offset = float(dict_["offset"])
+            instance.offset = int(dict_["offset"])
         elif include_empty:
-            instance.offset = float()
+            instance.offset = int()
         if "endDate" in dict_ and dict_["endDate"] is not None:
             instance.end_date = str(dict_["endDate"])
         elif include_empty:
             instance.end_date = str()
         if "pageSize" in dict_ and dict_["pageSize"] is not None:
-            instance.page_size = float(dict_["pageSize"])
+            instance.page_size = int(dict_["pageSize"])
         elif include_empty:
-            instance.page_size = float()
+            instance.page_size = int()
         if "startDate" in dict_ and dict_["startDate"] is not None:
             instance.start_date = str(dict_["startDate"])
         elif include_empty:
