+++ accelbyte_py_sdk/api/iam/operations/users/get_user_login_histories.py
@@ -65,7 +65,7 @@ class GetUserLoginHistories(Operation):
 
         before: (before) OPTIONAL float in query
 
-        limit: (limit) OPTIONAL float in query
+        limit: (limit) OPTIONAL int in query
 
     Responses:
         200: OK - ModelLoginHistoriesResponse (OK)
@@ -90,7 +90,7 @@ class GetUserLoginHistories(Operation):
     user_id: str                                                                                   # REQUIRED in [path]
     after: float                                                                                   # OPTIONAL in [query]
     before: float                                                                                  # OPTIONAL in [query]
-    limit: float                                                                                   # OPTIONAL in [query]
+    limit: int                                                                                     # OPTIONAL in [query]
 
     # endregion fields
 
@@ -198,7 +198,7 @@ class GetUserLoginHistories(Operation):
         self.before = value
         return self
 
-    def with_limit(self, value: float) -> GetUserLoginHistories:
+    def with_limit(self, value: int) -> GetUserLoginHistories:
         self.limit = value
         return self
 
@@ -225,9 +225,9 @@ class GetUserLoginHistories(Operation):
         elif include_empty:
             result["before"] = float()
         if hasattr(self, "limit") and self.limit:
-            result["limit"] = float(self.limit)
+            result["limit"] = int(self.limit)
         elif include_empty:
-            result["limit"] = float()
+            result["limit"] = int()
         return result
 
     # endregion to methods
@@ -279,7 +279,7 @@ class GetUserLoginHistories(Operation):
         user_id: str,
         after: Optional[float] = None,
         before: Optional[float] = None,
-        limit: Optional[float] = None,
+        limit: Optional[int] = None,
     ) -> GetUserLoginHistories:
         instance = cls()
         instance.namespace = namespace
@@ -312,9 +312,9 @@ class GetUserLoginHistories(Operation):
         elif include_empty:
             instance.before = float()
         if "limit" in dict_ and dict_["limit"] is not None:
-            instance.limit = float(dict_["limit"])
+            instance.limit = int(dict_["limit"])
         elif include_empty:
-            instance.limit = float()
+            instance.limit = int()
         return instance
 
     @staticmethod
